Added API explanation
One of our hosted users expressed confusion about the API tab on the User Profile and Michael suggested we add something to the documentation explaining what this is, to refer people there in the future (until this feature is hidden for most users per https://github.com/pkp/pkp-lib/issues/3635). Do you think this explanation is sufficient and accurate?