Renaming and restructuring user-accounts.md
Further discussion can be seen in the documentation Slack channel, but to accommodate the rewrite of the roles/permissions documentation started in the most recent DIG, a rehaul of this chapter is needed. All roles & permissions information will be moved to users-and-roles.md, with this chapter being dedicated to account creation/settings management. I suggest "Account Management" as a title (or maybe "Your Account"?) but I am open to suggestions.