Update with information about filters
To still do: replace the screencaps to show new filters. I couldn't get the "filter by editor" to work on the sandbox. Unclear if there needs to be more than 2 editors with assignments, or if the feature is glitchy. In any case, leaving the 3.2 screencaps for now, but will make a note in the spreadsheet.