Fixed broken link in Upgrading from OJS 2 to 3
The link was to a chapter in Learning OJS 3 called What's New in OJS 3. When the DIG updated Learning OJS 3 for OJS 3.3, they changed this chapter to "What's new in this version of OJS." I changed the link in the upgrading guide to the Learning OJS - 3.2 version, which still explains the differences between the OJS 2 and 3 interface.